pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
